+++ gatsby-browser.js
@@ -1,0 +1,3 @@
+
+require('./src/styles/base.scss');
+require('./src/styles/syntax-highlighter.scss');