+++ gatsby-config.js
@@ -1,0 +1,46 @@
+/**
+ * Configure your Gatsby site with this file.
+ *
+ * See: https://www.gatsbyjs.org/docs/gatsby-config/
+ */
+
+// Loads .env.development in order to enable Playground mode of Graphql
+require('dotenv').config({
+  path: `.env.${process.env.NODE_ENV}`,
+});
+
+module.exports = {
+  siteMetadata: {
+    title: 'Adriano Rosa',
+    description: `Analista Programador Backend e Frontend - Web Developer FullStack`,
+    author: 'Adriano Rosa',
+  },
+  // assetPrefix: `https://cdn.example.com`,
+  plugins: [
+    `gatsby-plugin-react-helmet`,
+    `gatsby-plugin-sass`,
+    {
+      resolve: `gatsby-source-filesystem`,
+      options: {
+        name: `posts`,
+        path: `${__dirname}/_source/blog/_posts/`,
+      },
+    },
+
+    {
+      resolve: `gatsby-transformer-remark`,
+      options: {
+        tableOfContents: {
+          heading: null,
+          maxDepth: 3,
+        },
+        plugins: [
+          {
+            resolve: `gatsby-remark-prismjs`,
+            options: {},
+          },
+        ],
+      },
+    },
+  ],
+};