+++ src/components/footer.js
@@ -1,0 +1,22 @@
+/** ========================================================================
+ * Project     : gatsby-markdown-starter
+ * Component   : footer
+ * Author      : Adriano Rosa <https://adrianorosa.com>
+ * Date        : 2019-11-09 15:31
+ * ========================================================================
+ * Copyright 2019 Adriano Rosa <https://adrianorosa.com>
+ * ======================================================================== */
+
+import React from 'react';
+
+const Footer = props => {
+  return (
+    <>
+      <footer>
+        <small>© adrianorosa.com</small>
+      </footer>
+    </>
+  );
+};
+
+export default Footer;